% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assing_aadt.R
\name{assign_aadt_major}
\alias{assign_aadt_major}
\title{Assign AADT data to road network}
\usage{
assign_aadt_major(
  lines,
  junctions,
  traffic,
  bounds,
  major_ref = c("motorway", "motorway_link", "primary", "primary_link", "trunk",
    "trunk_link")
)
}
\arguments{
\item{lines}{data frame of road network from `extract_from_osm`}

\item{junctions}{data frame of junction points from `extract_from_osm`}

\item{traffic}{AADT counts}

\item{bounds}{boundaries of study area}

\item{major_ref}{names of major road types}
}
\description{
Assign AADT data to road network
}
