% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_dft_aadt}
\alias{download_dft_aadt}
\title{Download DfT Traffic Counts}
\usage{
download_dft_aadt(
  url =
    "https://storage.googleapis.com/dft-statistics/road-traffic/downloads/data-gov-uk/dft_traffic_counts_aadf.zip"
)
}
\arguments{
\item{url}{URL of DfT data}
}
\description{
Download DfT Traffic Counts
}
