% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_lsoa}
\alias{download_lsoa}
\title{Download LSOA bounds}
\usage{
download_lsoa(
  url =
    "https://github.com/ITSLeeds/MinorRoadTraffic/releases/download/Input_Data/LSOA_2011.gpkg"
)
}
\arguments{
\item{url}{URL of data}
}
\description{
Download LSOA bounds
}
