% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_osm}
\alias{download_osm}
\title{Download OSM data for a named location}
\usage{
download_osm(x = "Leeds, UK", bbox = NULL)
}
\arguments{
\item{x}{name passed to osmdata default "Leeds, UK"}

\item{bbox}{alterative method using a bounding box}
}
\description{
Download OSM data for a named location
}
