% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_pbcc}
\alias{download_pbcc}
\title{Download PBCC data}
\usage{
download_pbcc(
  url =
    "https://github.com/creds2/CarbonCalculator/releases/download/1.0/PBCC_LSOA_data.zip"
)
}
\arguments{
\item{url}{URL of data}
}
\description{
Downloads data from www.carbon.place
}
