% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_osm.R
\name{extract_from_osm}
\alias{extract_from_osm}
\title{Extract road network from OSM}
\usage{
extract_from_osm(
  network,
  bounds,
  road_types = c("motorway", "motorway_link", "trunk", "trunk_link", "primary",
    "primary_link", "secondary", "secondary_link", "tertiary", "tertiary_link",
    "unclassified", "living_street", "residential"),
  col_names = c("osm_id", "name", "ref", "highway", "junction", "maxspeed", "geometry")
)
}
\arguments{
\item{network}{raw OSM lines}

\item{bounds}{bounding polygon from `make_convex_hull`}

\item{road_types}{road types to be kept}

\item{col_names}{column names to be kept}
}
\description{
Extract road network from OSM
}
