% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/road_density.R
\name{line_segment}
\alias{line_segment}
\title{Segments lines
From stplanr??}
\usage{
line_segment(l, segment_length = NA)
}
\arguments{
\item{l}{r?}

\item{segment_length}{?}
}
\description{
Segments lines
From stplanr??
}
