% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{make_convex_hull}
\alias{make_convex_hull}
\title{Make convex hull around traffic points}
\usage{
make_convex_hull(x, dist = 1000)
}
\arguments{
\item{x}{data frame of sf points}

\item{dist}{buffer distance, default 1000 metres}
}
\description{
Make convex hull around traffic points
}
