% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assing_aadt.R
\name{minor_roads_distance}
\alias{minor_roads_distance}
\title{Aim: To calculate AADT for minor roads
This script assumes the data has already been cleaned and the AADT for
major roads assigned
Rework of original file dropping the splitting an buffering stage}
\usage{
minor_roads_distance(lines, junc_majmi)
}
\arguments{
\item{lines}{data frame of road network from `extract_from_osm`}

\item{junc_majmi}{data frame of junctions between major and minor roads}
}
\description{
Aim: To calculate AADT for minor roads
This script assumes the data has already been cleaned and the AADT for
major roads assigned
Rework of original file dropping the splitting an buffering stage
}
