% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assing_aadt.R
\name{osm_fill_ref}
\alias{osm_fill_ref}
\title{Fill in missing road references}
\usage{
osm_fill_ref(
  lines,
  nearest_neighbour = 30,
  major_ref = c("motorway", "motorway_link", "primary", "primary_link", "trunk",
    "trunk_link")
)
}
\arguments{
\item{lines}{data frame of road network from `extract_from_osm`}

\item{nearest_neighbour}{number of nearest neighbours to search for, default
= 30}

\item{major_ref}{names of major road types}
}
\description{
Sometimes OSM roads are missing the ref value (e.g. M25). This fills in the
missing values based on nearest neighbours.
}
