% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assing_aadt.R
\name{osm_fill_ref2}
\alias{osm_fill_ref2}
\title{Fill in missing road references}
\usage{
osm_fill_ref2(
  lines,
  major_ref = c("motorway", "motorway_link", "primary", "primary_link", "trunk",
    "trunk_link"),
  max_iter = 10
)
}
\arguments{
\item{lines}{data frame of road network from `extract_from_osm`}

\item{major_ref}{names of major road types}

\item{max_iter}{number of iterations (default 10)}
}
\description{
Sometimes OSM roads are missing the ref and name value (e.g. M25). This fills
in the missing values based on the adjacent links by an iterative process.
}
