% Generated by manual addition
% Please edit documentation in R/predict_minor_flow_rf.R
\name{predict_minor_flow_rf}
\alias{predict_minor_flow_rf}
\title{Predict minor road traffic using random forest}
\usage{
predict_minor_flow_rf(network, seed = 42)
}
\arguments{
  \item{network}{sf object containing the road network. Should contain the
  columns described in \code{\link{predict_minor_flow_rf}}.}
  \item{seed}{integer used to initialise the random number generator.}
}
\value{
An sf object identical to \code{network} with an additional column
\code{rf_prediction}. The fitted model is attached as an attribute
named \code{"rf_model"}.
}
\description{
Trains a random forest regression model using the \pkg{ranger} package and
predicts minor road traffic for all links in the network.
}
\examples{
\dontrun{
net <- predict_minor_flow_rf(net)
}
}
