% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/road_density.R
\name{road_density}
\alias{road_density}
\title{Calculate road density #'}
\usage{
road_density(
  lines,
  zones,
  segment_length = 1000,
  use_centroids = FALSE,
  zones_id = names(zones)[1]
)
}
\arguments{
\item{lines}{road network}

\item{zones}{zones for stats}

\item{segment_length}{max length of a road segment in metres}

\item{use_centroids}{logical if true road centroids are used, faster but less
accurate}

\item{zones_id}{character, name of the column in `zones` that has the zone
ID, default to `names(zones)[1]`}
}
\description{
Calculate road density #'
}
