% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/road_centrality.R
\name{road_centrality}
\alias{road_centrality}
\title{Calculate centrality of minor roads}
\usage{
road_centrality(lines, threshold = 5)
}
\arguments{
\item{lines}{road network}

\item{threshold}{minimum number of minor roads to include in a zone, default 5}
}
\description{
Calculate centrality of minor roads
}
